refactor: group index-only scan implementation
